Add IPv6 support (#44)
In order to support IPv6 on AWS we had to do a major rework of the
networking configuration of the VPC and ECS clusters.

In case you like to enjoy some some trivia about why IPv6 support is so
weak in AWS, check these links:
* https://news.ycombinator.com/item?id=37608900
* https://github.com/DuckbillGroup/aws-ipv6-gaps
* https://github.com/aws/containers-roadmap/issues/1340

More details follow:
* ECS tasks network mode has been switched to awsvpc, which is the
[official way to support
IPv6](https://aws.amazon.com/about-aws/whats-new/2020/11/amazon-ecs-supports-ipv6-in-awsvpc-networking-mode/).
* Separate the networking stack into private and public. The routing
table of the public network has direct IPv4 and IPv6 routes configured,
while the private network uses NAT to route to the internet.
* For NAT to work we also need to allocate elastic IPs, one for each
availability zone we would like to support.

In order to get the networking to work properly in ECS, we setup the
container host to make use of the public network to have direct access
to the internet, while the ECS container itself uses the private NATed
network. This is the desired configuration because we don't want to
directly expose the container to the internet, but rather map the
container port to the public internet accessible endpoint via load
balancer.