Fix bug in availability zone mapping (#49)
Previously we had a bug in the netblock arithmetics for allocating the
private vs the public subnets based on availability zones. Due to this
bug we had instances being deployed into private networks which had an
availability zone that wasn't in our list of supported zones and
therefore we were never deploying tasks onto it.

In order to address this, since it requires changing the networking
stack, we had to also implement a lot of changes to how security groups
are being created and switched, which was anyways a good things to do.

The key point is that when you would like to modify a security group,
because for example you need to change it's subnet, you need to create
the new security group before deleting the old one.
In order to do this, you need to make use of `name_prefix` instead of
static `name` attributes.

This PR does that.