fix(prod): increase asg min capacity (#106)
`api.ooni.org` gives us a `503` when calling the incidents service. On
further inspection, it seems the redeployment of the findings service
has been halted due to insufficient memory in the ecs infra setup. This
diff increase the asg capacity to a minimum of 4 ec2 instances which
should suffice for now.